feat: delete components from FormVacancies and implements components from Commons
