+++ src/components/FormVacancies/styles.js
@@ -1,11 +1,7 @@
 import styled from 'styled-components';
+
 import Button from '@mui/material/Button';
-import { FormSwitch } from './FormSwitch';
 
 export const FormButton = styled(Button)`
-    display: ${(props) => props.display};
-`;
-
-export const FormEditSwitch = styled(FormSwitch)`
     display: ${(props) => props.display};
 `;
